+++ packages/std/src/results/context.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use schemars::JsonSchema;
 use std::fmt;
 
@@ -9,6 +11,10 @@ use super::handle::HandleResponse;
 use super::init::InitResponse;
 use super::migrate::MigrateResponse;
 
+#[deprecated(
+    since = "0.14.0",
+    note = "Import the trait cosmwasm_std::MutResponse in order to use mutating helpers on InitResponse/HandleResponse/MigrateResponse directly."
+)]
 #[derive(Clone, Debug, PartialEq)]
 pub struct Context<T = Empty>
 where