+++ CHANGELOG.md
@@ -76,6 +76,12 @@ and this project adheres to
 [#736]: https://github.com/CosmWasm/cosmwasm/pull/736
 [#690]: https://github.com/CosmWasm/cosmwasm/issues/690
 
+### Deprecated
+
+- cosmwasm-std: `Context` is deprecated in favour of the new `MutResponse` trait
+  that adds mutable helpers to `InitResponse`, `MigrateResponse` and
+  `HandleResponse` directly.
+
 ## [0.13.2] - 2021-01-14
 
 ## Changed